+++ package.json
@@ -15,7 +15,7 @@
     "react-icons": "^4.7.1",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.4.3",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "reactstrap": "^9.1.5",
     "remixicon": "^2.5.0",
     "sweetalert2": "^11.6.10",
