+++ src/routers/PrivateRoute.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const PrivateRoute = () => {
   return (
-    <div>s</div>
+    <div>PrivateRoute</div>
   );
 };
